[EDQ-401] Renamed the 'table_metrics' scripts to be more informative.
