+++ data_steward/analytics/table_metrics/weekly_scripts/drug_route_success_rate.py
@@ -222,248 +222,6 @@ site_df = pd.merge(site_map, site_df, how='outer', on='src_hpo_id')
 
 site_df
 
-Lipid = (40782589, 40795800, 40772572)
-
-CBC = (40789356, 40789120, 40789179, 40772748, 40782735, 40789182, 40786033,
-       40779159)
-
-CBCwDiff = (40785788, 40785796, 40779195, 40795733, 40795725, 40772531,
-            40779190, 40785793, 40779191, 40782561, 40789266)
-
-CMP = (3049187, 3053283, 40775801, 40779224, 40782562, 40782579, 40785850,
-       40785861, 40785869, 40789180, 40789190, 40789527, 40791227, 40792413,
-       40792440, 40795730, 40795740, 40795754)
-
-Physical_Measurement = (40654163, 40655804, 40654162, 40655805, 40654167,
-                        40654164)
-
-measurement_codes = Lipid + CBC + CBCwDiff + CMP + Physical_Measurement
-
-# # Integration of Units for All Measurements:
-#
-
-# #### Getting the numbers for all of the unit concept IDs by site
-
-# +
-unit_concept_ids_by_site_query = """
-CREATE TABLE `{DATASET}.sites_unit_counts`
-OPTIONS (
-expiration_timestamp=TIMESTAMP_ADD(CURRENT_TIMESTAMP(), INTERVAL 3 MINUTE)
-)
-AS
-SELECT
-DISTINCT
-mm.src_hpo_id, COUNT(m.measurement_id) as number_total_rows
-FROM
-`{DATASET}.unioned_ehr_measurement` m
-JOIN
-`{DATASET}._mapping_measurement` mm
-ON
-m.measurement_id = mm.measurement_id 
-JOIN
-`{DATASET}.concept` c
-ON
-m.unit_concept_id = c.concept_id 
-GROUP BY 1
-ORDER BY number_total_rows DESC
-""".format(DATASET = DATASET)
-
-unit_concept_ids_by_site = pd.io.gbq.read_gbq(unit_concept_ids_by_site_query, dialect='standard')
-
-# +
-unit_concept_ids_by_site_query = """
-SELECT
-*
-FROM
-`{DATASET}.sites_unit_counts`
-""".format(DATASET = DATASET)
-
-unit_concept_ids_by_site = pd.io.gbq.read_gbq(unit_concept_ids_by_site_query, dialect='standard')
-# -
-
-unit_concept_ids_by_site
-
-# #### Below are the "successful" unit concept IDs
-
-# +
-successful_unit_concept_ids_by_site_query = """
-CREATE TABLE `{DATASET}.sites_successful_unit_counts`
-OPTIONS (
-expiration_timestamp=TIMESTAMP_ADD(CURRENT_TIMESTAMP(), INTERVAL 3 MINUTE)
-)
-AS
-SELECT
-DISTINCT
-mm.src_hpo_id, COUNT(m.measurement_id) as number_valid_units
-FROM
-`{DATASET}.unioned_ehr_measurement` m
-JOIN
-`{DATASET}._mapping_measurement` mm
-ON
-m.measurement_id = mm.measurement_id 
-JOIN
-`{DATASET}.concept` c
-ON
-m.unit_concept_id = c.concept_id 
-WHERE
-c.standard_concept IN ('S')
-AND
-LOWER(c.domain_id) LIKE '%unit%'
-GROUP BY 1
-ORDER BY number_valid_units DESC
-""".format(DATASET = DATASET)
-
-successful_unit_concept_ids_by_site = pd.io.gbq.read_gbq(successful_unit_concept_ids_by_site_query, dialect='standard')
-
-# +
-successful_unit_concept_ids_by_site_query = """
-SELECT
-*
-FROM
-`{DATASET}.sites_successful_unit_counts`
-""".format(DATASET = DATASET)
-
-successful_unit_concept_ids_by_site = pd.io.gbq.read_gbq(successful_unit_concept_ids_by_site_query, dialect='standard')
-# -
-
-successful_unit_concept_ids_by_site
-
-final_all_units_df = pd.merge(site_df, unit_concept_ids_by_site, on = 'src_hpo_id', how = 'left')
-
-final_all_units_df = pd.merge(final_all_units_df, successful_unit_concept_ids_by_site, on = 'src_hpo_id', how = 'left')
-
-final_all_units_df['total_unit_success_rate'] = round(final_all_units_df['number_valid_units'] / final_all_units_df['number_total_rows'] * 100, 2)
-
-final_all_units_df = final_all_units_df.fillna(0)
-
-final_all_units_df = final_all_units_df.sort_values(by='total_unit_success_rate', ascending = False)
-
-final_all_units_df
-
-# # Integration of Units for Selected Measurements
-#
-# #### making the distinction because - according to the [AoU EHR Operations](https://sites.google.com/view/ehrupload/omop-tables/measurement?authuser=0)  page (as of 03/11/2020) - the unit_concept_id are only required for the 'required labs'
-
-# +
-selected_unit_concept_ids_by_site_query = """
-CREATE TABLE `{DATASET}.sites_unit_counts_selected_measurements`
-OPTIONS (
-expiration_timestamp=TIMESTAMP_ADD(CURRENT_TIMESTAMP(), INTERVAL 3 MINUTE)
-)
-AS
-SELECT
-DISTINCT
-mm.src_hpo_id, COUNT(m.measurement_id) as number_total_units_selected_measurements
-FROM
-`{DATASET}.unioned_ehr_measurement` m
-JOIN
-`{DATASET}._mapping_measurement` mm
-ON
-m.measurement_id = mm.measurement_id 
-JOIN
-`{DATASET}.concept_ancestor` ca
-ON
-ca.descendant_concept_id = m.measurement_concept_id
-JOIN
-`{DATASET}.concept` c
-ON
-m.unit_concept_id = c.concept_id
-WHERE
-ca.ancestor_concept_id IN {selected_measurements}
-GROUP BY 1
-ORDER BY number_total_units_selected_measurements DESC
-""".format(DATASET = DATASET, selected_measurements = measurement_codes)
-
-selected_unit_concept_ids_by_site = pd.io.gbq.read_gbq(selected_unit_concept_ids_by_site_query, dialect='standard')
-
-# +
-selected_unit_concept_ids_by_site_query = """
-SELECT
-*
-FROM
-`{DATASET}.sites_unit_counts_selected_measurements`
-""".format(DATASET = DATASET)
-
-selected_unit_concept_ids_by_site = pd.io.gbq.read_gbq(selected_unit_concept_ids_by_site_query, dialect='standard')
-# -
-
-selected_unit_concept_ids_by_site
-
-# #### Below are the 'successful' unit_concept_ids
-
-# +
-successful_selected_unit_concept_ids_by_site_query = """
-CREATE TABLE `{DATASET}.sites_successful_unit_counts_selected_measurements`
-OPTIONS (
-expiration_timestamp=TIMESTAMP_ADD(CURRENT_TIMESTAMP(), INTERVAL 3 MINUTE)
-)
-AS
-SELECT
-DISTINCT
-mm.src_hpo_id, COUNT(m.unit_concept_id) as number_valid_units_selected_measurements
-FROM
-`{DATASET}.unioned_ehr_measurement` m
-JOIN
-`{DATASET}._mapping_measurement` mm
-ON
-m.measurement_id = mm.measurement_id 
-JOIN
-`{DATASET}.concept_ancestor` ca
-ON
-ca.descendant_concept_id = m.measurement_concept_id
-JOIN
-`{DATASET}.concept` c
-ON
-m.unit_concept_id = c.concept_id
-WHERE
-c.standard_concept IN ('S')
-AND
-LOWER(c.domain_id) LIKE '%unit%'
-AND
-ca.ancestor_concept_id IN {selected_measurements}
-GROUP BY 1
-ORDER BY number_valid_units_selected_measurements DESC
-""".format(DATASET = DATASET, selected_measurements = measurement_codes)
-
-successful_selected_unit_concept_ids_by_site = pd.io.gbq.read_gbq(successful_selected_unit_concept_ids_by_site_query, dialect='standard')
-
-# +
-successful_selected_unit_concept_ids_by_site_query = """
-SELECT
-*
-FROM
-`{DATASET}.sites_successful_unit_counts_selected_measurements`
-""".format(DATASET = DATASET)
-
-successful_selected_unit_concept_ids_by_site = pd.io.gbq.read_gbq(successful_selected_unit_concept_ids_by_site_query, dialect='standard')
-# -
-
-successful_selected_unit_concept_ids_by_site
-
-# +
-final_all_units_df = pd.merge(final_all_units_df, selected_unit_concept_ids_by_site, on = 'src_hpo_id', how = 'left')
-
-final_all_units_df = pd.merge(final_all_units_df, successful_selected_unit_concept_ids_by_site, on = 'src_hpo_id', how = 'left')
-# -
-
-final_all_units_df
-
-final_all_units_df['total_proportion_of_units_that_are_selected'] = round(final_all_units_df['number_total_rows'] / final_all_units_df['number_total_units_selected_measurements'] * 100, 2)
-
-
-final_all_units_df['selected_unit_success_rate'] = round(final_all_units_df['number_valid_units_selected_measurements'] / final_all_units_df['number_total_units_selected_measurements'] * 100, 2)
-
-
-# +
-final_all_units_df = final_all_units_df.fillna(0)
-
-final_all_units_df = final_all_units_df.sort_values(by='total_unit_success_rate', ascending = False)
-
-final_all_units_df
-# -
-
-final_all_units_df.to_csv("{cwd}\measurement_units.csv".format(cwd = cwd))
-
 # #  Integration of Routes for Select Drugs:
 # #### This is required for all drugs according to the [AoU EHR Operations Page](https://sites.google.com/view/ehrupload/omop-tables/drug_exposure?authuser=0)
