+++ data_steward/analytics/table_metrics/weekly_scripts/measurement_integration_success_rate.py
